01/11

+++ src/main/java/tn/esprit/spring1/tpyasmine/Repository/AbonnementRepository.java
@@ -2,5 +2,5 @@ package tn.esprit.spring1.tpyasmine.Repository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.spring1.tpyasmine.entities.Abonnement;
 
-public interface AbonnementRepository extends CrudRepository<Abonnement, Integer>{
+public interface AbonnementRepository extends CrudRepository<Abonnement, Long>{
 }
