+++ src/main/java/tn/esprit/spring1/tpyasmine/Repository/CoursRepository.java
@@ -2,5 +2,5 @@ package tn.esprit.spring1.tpyasmine.Repository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.spring1.tpyasmine.entities.Cours;
 
-public interface CoursRepository extends CrudRepository<Cours, Integer>{
+public interface CoursRepository extends CrudRepository<Cours, Long>{
 }
