+++ src/main/java/tn/esprit/spring1/tpyasmine/Repository/InscriptionRepository.java
@@ -2,5 +2,5 @@ package tn.esprit.spring1.tpyasmine.Repository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.spring1.tpyasmine.entities.Inscription;
 
-public interface InscriptionRepository extends CrudRepository<Inscription, Integer>{
+public interface InscriptionRepository extends CrudRepository<Inscription, Long>{
 }
