+++ src/main/java/tn/esprit/spring1/tpyasmine/Repository/MoniteurRepository.java
@@ -2,5 +2,5 @@ package tn.esprit.spring1.tpyasmine.Repository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.spring1.tpyasmine.entities.Moniteur;
 
-public interface MoniteurRepository extends CrudRepository<Moniteur, Integer>{
+public interface MoniteurRepository extends CrudRepository<Moniteur, Long>{
 }
