+++ src/main/java/tn/esprit/spring1/tpyasmine/Repository/PisteRepository.java
@@ -2,5 +2,5 @@ package tn.esprit.spring1.tpyasmine.Repository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.spring1.tpyasmine.entities.Piste;
 
-public interface PisteRepository extends CrudRepository<Piste, Integer>{
+public interface PisteRepository extends CrudRepository<Piste, Long>{
 }
