+++ src/main/java/tn/esprit/spring1/tpyasmine/Repository/SkieurRepository.java
@@ -3,6 +3,6 @@ package tn.esprit.spring1.tpyasmine.Repository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.spring1.tpyasmine.entities.Skieur;
 
-public interface SkieurRepository extends CrudRepository<Skieur, Integer>{
+public interface SkieurRepository extends CrudRepository<Skieur, Long>{
 
 }
